# docs/swagger.yaml
basePath: /
definitions:
...
  handler.CreateAssessmentRequest:
    properties:
      grip_position:
        type: integer
      left_value:
        type: number
      right_value:
        type: number
      session_id:
        type: string
      type:
        type: integer
    type: object
  handler.CreateBuiltinTrainingWeightRequest:
    properties:
      builtin_training_id:
        type: string
      custom_weight_left:
        type: number
      custom_weight_right:
        type: number
      id:
        type: string
    type: object
...
  handler.UpdateBuiltinTrainingWeightRequest:
    properties:
      custom_weight_left:
        type: number
      custom_weight_right:
        type: number
    type: object
...
host: api.crimpy.app
info:
  contact:
    email: support@crimpy.com
    name: API Support
  description: Backend API for Crimpy climbing training application
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://api.crimpy.app/terms
  title: Crimpy API
  version: "1.0"
paths:
...
  /api/assessments:
    get:
      description: Retrieve all assessment results for the authenticated user, ordered
        by session date descending
      produces:
      - application/json
      responses:
        "200":
          description: List of assessments
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all assessments for the current user
      tags:
      - Assessment
    post:
      consumes:
      - application/json
      description: Create a new assessment result for the authenticated user, linked
        to an existing session they own
      parameters:
      - description: Assessment details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAssessmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Assessment created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Session not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create an assessment linked to an existing session
      tags:
      - Assessment
  /api/assessments/{id}:
    delete:
      description: Delete an assessment. User must own the associated session.
      parameters:
      - description: Assessment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an assessment
      tags:
      - Assessment
...
  /api/builtin-training-weights:
    get:
      consumes:
      - application/json
      description: Retrieve all custom weight overrides for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of weight overrides
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all builtin training weight overrides
      tags:
      - BuiltinTrainingWeight
    post:
      consumes:
      - application/json
      description: Create a custom weight override for a builtin training
      parameters:
      - description: Weight override details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateBuiltinTrainingWeightRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Weight override created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a builtin training weight override
      tags:
      - BuiltinTrainingWeight
  /api/builtin-training-weights/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a custom weight override. User must own it.
      parameters:
      - description: Weight override ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a builtin training weight override
      tags:
      - BuiltinTrainingWeight
    put:
      consumes:
      - application/json
      description: Update custom weights. User must own it.
      parameters:
      - description: Weight override ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated weights
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateBuiltinTrainingWeightRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated weight override
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a builtin training weight override
      tags:
      - BuiltinTrainingWeight
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"